remove canCustomErrorService. Rename error constants file
